시바충돌체 관심사 분리 및, 미들웨어 jpg검사생략, dotnaming으로 depth줄이기
