+++ features/luckyshiba/components/background/collision.type.ts
