+++ middleware.ts
@@ -4,5 +4,5 @@ import { authConfig } from './auth';
 export default NextAuth(authConfig).auth;
 
 export const config = {
-  matcher: ['/((?!api|_next/static|_next/image|.*\\.png$|.*\\.glb$).*)'],
+  matcher: ['/((?!api|_next/static|_next/image|.*.png$|.*.glb$|.*.jpg$).*)/'],
 };
